Add support for downloading input files from URLs
That is, a ROM resource (ROM file, CHD, zip file containing ROMs, etc.)
can be downloaded from some remote site and unpacked directly
from this program's memory without any intermediate steps.

Of course, since it works from RAM, it's intended for resources
that are relatively small (or systems with lots of RAM).

I'll probably drop in temp file support for very large files later
in order to avoid using up all the RAM unexpectedly,
but this is easier to implement for now.