Save specified ROM directories as defaults
That is, when doing a "emuman mame verify -r <some_dir>"
that <some_dir> will be saved in a config file so that
one can skip it on subsequent calls to the "emuman mame" functions.

This is a keystroke saver in the short term.  But in the long run,
it should enable me to have "redump verify-all" and "nointro verify-all"
routines by cross-referencing defined directories against
those in the cached databases.