Unify file and URL inputs and make them un-flagged options
That is, resources that look like URLs are treated as URLs,
and everything else is treated as a local file/directory.

Since these are common options that can be provided many times
(as opposed to the ROMs output directory, which is provided optionally once),
not having to flag each one is more convenient and works better
with scripts.